Task 32 rb

See merge request 2020ss-SEPM-group/ss20_sepm_inso_01!146

+++ backend/src/main/java/at/ac/tuwien/sepm/groupphase/backend/endpoint/dto/DetailedEventDto.java
@@ -26,7 +26,7 @@ public class DetailedEventDto {
     private List<SimpleShowDto> shows;
     private String type;
     private String category;
-    private List<Integer> prices;
+    private List<Double> prices;
     private int totalTicketsSold;
     private long duration;
 }
